etc/dists/scripts/*-set_ugid_quota.sh: enhancements
Rework these scripts to:

1. Use cat and shell's "here document" instead of echo

2. Remove error message printing ("Can't create script file") since
   it is now printed by cat itself.

3. Use awk to prevent changing /etc/mtab if it is already fixed for
   quota on case (newer kernel do that).

4. Try to get device name from the file, if not found revert to supplied one.

Inspired by commit e324b0ade (which, well, was mostly right).

Somewhat relevant to bug #1904:
http://bugzilla.openvz.org/1904

+++ etc/dists/scripts/arch-set_ugid_quota.sh
@@ -29,22 +29,20 @@ if [ -z "$MAJOR" ]; then
 	exit 0
 fi
 
-echo '#!/bin/sh
+cat << EOF > ${SCRIPTANAME} || exit 1
+#!/bin/sh
 
-[ -e "/dev/'${DEVFS}'" ] || mknod /dev/'${DEVFS}' b '$MAJOR' '$MINOR'
-rm -f /etc/mtab >/dev/null 2>&1
-
-echo "/dev/'${DEVFS}' / reiserfs rw,usrquota,grpquota 0 0" > /etc/mtab
-mnt=`grep -v " / " /proc/mounts`
-if [ $? = 0 ]; then
-	echo "$mnt" >> /etc/mtab
+dev=\$(awk '(\$2 == "/") && (\$4 ~ /usrquota/) && (\$4 ~ /grpquota/) {print \$1}' /etc/mtab)
+if test -z "\$dev"; then
+	dev="/dev/${DEVFS}"
+	rm -f /etc/mtab >/dev/null 2>&1
+	echo "/dev/${DEVFS} / reiserfs rw,usrquota,grpquota 0 0" > /etc/mtab
+	grep -v " / " /proc/mounts >> /etc/mtab 2>/dev/null
+	chmod 644 /etc/mtab
 fi
-chmod 644 /etc/mtab
+[ -e "\$dev" ] || mknod \$dev b $MAJOR $MINOR
 quotaon -aug
-' > ${SCRIPTANAME} || {
-	echo "Unable to create ${SCRIPTANAME}"
-	exit 1
-}
+EOF
 chmod 755 ${SCRIPTANAME}
 
 # add vzquota to rc.conf
