+++ etc/dists/scripts/slackware-set_ugid_quota.sh
@@ -27,16 +27,18 @@ if [ -z "$MAJOR" ]; then
 	ln -sf /proc/mounts /etc/mtab
 	exit 0
 fi
-echo '#!/bin/sh
+
+cat << EOF > ${SCRIPTANAME} || exit 1
 start() {
-	[ -e "/dev/'${DEVFS}'" ] || mknod /dev/'${DEVFS}' b '$MAJOR' '$MINOR'
-	rm -f /etc/mtab >/dev/null 2>&1
-	echo "/dev/'${DEVFS}' / reiserfs rw,usrquota,grpquota 0 0" > /etc/mtab
-	mnt=`grep -v " / " /proc/mounts`
-	if [ $? = 0 ]; then
-		echo "$mnt" >> /etc/mtab
+	dev=\$(awk '(\$2 == "/") && (\$4 ~ /usrquota/) && (\$4 ~ /grpquota/) {print \$1}' /etc/mtab)
+	if test -z "\$dev"; then
+		dev="/dev/${DEVFS}"
+		rm -f /etc/mtab >/dev/null 2>&1
+		echo "/dev/${DEVFS} / reiserfs rw,usrquota,grpquota 0 0" > /etc/mtab
+		grep -v " / " /proc/mounts >> /etc/mtab 2>/dev/null
+		chmod 644 /etc/mtab
 	fi
-	chmod 644 /etc/mtab
+	[ -e "\$dev" ] || mknod \$dev b $MAJOR $MINOR
 	quotaon -aug
 }
 case "$1" in
@@ -45,10 +47,8 @@ case "$1" in
 	;;
   *)
 	exit
-esac ' > ${SCRIPTANAME} || {
-	echo "Unable to create ${SCRIPTANAME}"
-	exit 1
-}
+esac
+EOF
 chmod 755 ${SCRIPTANAME}
 
 if ! grep -q "${SCRIPTANAME}" ${RC_LOCAL} 2>/dev/null; then
